Extract module item dispatch helper
