+++ src/frontend/parser.ts
@@ -9,28 +9,15 @@ import { canonicalModuleId } from '../moduleIdentity.js';
 import { buildLogicalLines, getLogicalLine, type LogicalLine } from './parseLogicalLines.js';
 import {
   createModuleItemDispatchTable,
+  dispatchModuleItem,
   type ParseItemContext,
   type ParseItemResult,
 } from './parseModuleItemDispatch.js';
-import { topLevelStartKeyword } from './parseModuleCommon.js';
-import {
-  parseExportModifier,
-  recoverUnsupportedParserLine,
-} from './parseParserRecovery.js';
-import {
-  looksLikeRawDataDirectiveStart,
-  maybeCloseSection,
-  parseSectionBodyItem,
-  parseSectionItems as parseSectionItemsFromHelper,
-} from './parseSectionBodies.js';
+import { parseSectionItems as parseSectionItemsFromHelper } from './parseSectionBodies.js';
 import { parseSectionHeader } from './parseSectionHeader.js';
 import { parseOpParamsFromText, parseParamsFromText } from './parseParams.js';
-import {
-  isReservedTopLevelDeclName,
-  stripLineComment as stripComment,
-} from './parseParserShared.js';
+import { isReservedTopLevelDeclName } from './parseParserShared.js';
 import { makeSourceFile, span, type SourceFile } from './source.js';
-import { parseDiag as diag } from './parseDiagnostics.js';
 
 /**
  * Parse a single `.zax` module file from an in-memory source string.
@@ -111,79 +98,14 @@ export function parseModuleFile(
   });
 
   function parseModuleItem(index: number, ctx: ParseItemContext): ParseItemResult {
-    const { raw, startOffset: lineStartOffset, endOffset: lineEndOffset } = getRawLine(index);
-    const text = stripComment(raw).trim();
-    const lineNo = logicalLines[index]?.lineNo ?? index + 1;
-    const filePath = logicalLines[index]?.filePath ?? modulePath;
-
-    if (text.length === 0) return { nextIndex: index + 1 };
-
-    if (ctx.scope === 'section') {
-      const sectionClose = maybeCloseSection(index, text, ctx, diagnostics);
-      if (sectionClose) return sectionClose;
-    }
-
-    const exportParsed = parseExportModifier({
-      text,
-      lineNo,
-      allowAsmSpecialCase: ctx.scope === 'module',
-      filePath,
+    return dispatchModuleItem(index, ctx, {
       diagnostics,
-    });
-    if (!exportParsed) return { nextIndex: index + 1 };
-
-    const hasExportPrefix = exportParsed.exported;
-    const rest = exportParsed.rest;
-    const stmtSpan = span(file, lineStartOffset, lineEndOffset);
-
-    if (ctx.scope === 'section') {
-      const parsedSectionItem = parseSectionBodyItem({
-        index,
-        ctx,
-        rest,
-        lineNo,
-        filePath,
-        stmtSpan,
-        diagnostics,
-      });
-      if (parsedSectionItem) return parsedSectionItem;
-    } else if (looksLikeRawDataDirectiveStart(rest)) {
-      diag(
-        diagnostics,
-        filePath,
-        `Raw data directives are only permitted inside data sections.`,
-        { line: lineNo, column: 1 },
-      );
-      return { nextIndex: index + 1 };
-    }
-
-    const dispatchKeyword = topLevelStartKeyword(rest);
-    const dispatchHandler =
-      dispatchKeyword === undefined ? undefined : moduleItemDispatchTable[dispatchKeyword];
-    if (dispatchHandler) {
-      const parsed = dispatchHandler({
-        index,
-        lineNo,
-        filePath,
-        text,
-        rest,
-        stmtSpan,
-        lineStartOffset,
-        hasExportPrefix,
-        ctx,
-      });
-      if (parsed) return parsed;
-    }
-
-    return recoverUnsupportedParserLine({
-      index,
-      scope: ctx.scope,
-      text,
-      rest,
-      hasExportPrefix,
-      lineNo,
-      filePath,
-      diagnostics,
+      file,
+      getRawLine,
+      logicalLines,
+      moduleItemDispatchTable,
+      modulePath,
+      span,
     });
   }
